+++ Solutions/GoogleWorkspaceReports/Data/Solution_GoogleWorkspaceReports.json
@@ -1,0 +1,46 @@
+{
+  "Name": "GoogleWorkspaceReports",
+  "Author": "Microsoft - support@microsoft.com",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Logos/google_logo.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [Google Workspace](https://workspace.google.com/) solution for Microsoft Sentinel enables you to ingest Google Workspace Activity events into Microsoft Sentinel. \n \n **Underlying Microsoft Technologies used:**\n\n This solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs: \n \n a. [Azure Monitor HTTP Data Collector API](https://docs.microsoft.com/azure/azure-monitor/logs/data-collector-api) \n \n b. [Azure Functions](https://azure.microsoft.com/services/functions/#overview)",
+  "Workbooks": [ 
+    "Workbooks/GoogleWorkspace.json" 
+   ],
+  "Parsers": [
+    "Parsers/GWorkspaceActivityReports"
+  ],
+  "Analytic Rules": [
+    "Analytic Rules/GWorkspaceAdminPermissionsGranted.yaml",
+    "Analytic Rules/GWorkspaceAlertEvents.yaml",
+    "Analytic Rules/GWorkspaceApiAccessToNewClient.yaml",
+    "Analytic Rules/GWorkspaceChangedUserAccess.yaml",
+    "Analytic Rules/GWorkspaceDifferentUAsFromSingleIP.yaml",
+    "Analytic Rules/GWorkspaceOutboundRelayAddedToSuiteDomain.yaml",
+    "Analytic Rules/GWorkspacePossibleBruteForce.yaml",
+    "Analytic Rules/GWorkspacePossibleMaldocFileNamesInGDRIVE.yaml",
+    "Analytic Rules/GWorkspaceTwoStepAuthenticationDisabledForUser.yaml",
+    "Analytic Rules/GWorkspaceUnexpectedOSUpdate.yaml"
+  ],
+  "Data Connectors": [
+    "Data Connectors/GWorkspaceReports_API_FunctionApp.json"
+  ],
+  "Hunting Queries": [
+    "Hunting Queries/GWorkspaceDocumentSharedExternally.yaml",
+    "Hunting Queries/GWorkspaceDocumentSharedPublicily.yaml",
+    "Hunting Queries/GWorkspaceDocumentSharedPublicilyWithLink.yaml",
+    "Hunting Queries/GWorkspaceMultiIPAddresses.yaml",
+    "Hunting Queries/GWorkspacePossibleSCAMSPAMorPhishingCalendar.yaml",
+    "Hunting Queries/GWorkspaceRareDocType.yaml",
+    "Hunting Queries/GWorkspaceSharedPrivateDocument.yaml",
+    "Hunting Queries/GWorkspaceSuspendedUsers.yaml",
+    "Hunting Queries/GWorkspaceUncommonUAsString.yaml",
+    "Hunting Queries/GWorkspaceUnknownLoginType.yaml",
+	"Hunting Queries/GWorkspaceUserReportedCalendarInviteAsSpam.yaml",
+    "Hunting Queries/GWorkspaceUserWithSeveralDevices.yaml"
+  ], 
+  "BasePath": "C:\\GitHub\\Azure-Sentinel\\Solutions\\GoogleWorkspaceReports",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+  "TemplateSpec": true,
+  "Is1Pconnector": false
+}
