+++ Solutions/GoogleWorkspaceReports/SolutionMetadata.json
@@ -1,0 +1,15 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-googleworkspacereports",
+	"firstPublishDate": "2022-01-24",
+	"providers": ["Google"],
+	"categories": {
+		"domains" : ["IT Operations"]
+	},
+	"support": {
+		"tier": "Microsoft",
+		"name": "Microsoft Corporation",
+		"email": "support@microsoft.com",
+		"link": "https://support.microsoft.com/"
+	}
+}
